Willow-willow-v5.1-gui-edition/willow_v5_1/requirements.txt
openai
google-generativeai
python-dotenv
gTTS
rich
PyQt6
